fix : Board DTO 구조 수정
