--- src/main/java/com/ureca/picky_be/base/business/board/dto/AddBoardContentReq.java
@@ -1,6 +1,0 @@
-package com.ureca.picky_be.base.business.board.dto;
-
-import com.ureca.picky_be.jpa.board.BoardContentType;
-
-public record AddBoardContentReq(String contentUrl, String type) {
-}
