--- src/main/java/com/ureca/picky_be/base/business/board/dto/AddBoardResp.java
@@ -1,5 +1,0 @@
-package com.ureca.picky_be.base.business.board.dto;
-
-public record AddBoardResp() {
-
-}
