--- src/main/java/com/ureca/picky_be/base/business/board/dto/GetAllBoardComments.java
@@ -1,4 +1,0 @@
-package com.ureca.picky_be.base.business.board.dto;
-
-public record GetAllBoardComments() {
-}
