--- src/main/java/com/ureca/picky_be/base/business/board/dto/UpdateBoardReq.java
@@ -1,6 +1,0 @@
-package com.ureca.picky_be.base.business.board.dto;
-
-import java.util.List;
-
-public record UpdateBoardReq(String boardContext, boolean isSpoiler) {
-}
