+++ src/main/java/com/ureca/picky_be/base/business/board/dto/boardDto/AddBoardReq.java
@@ -1,4 +1,6 @@
-package com.ureca.picky_be.base.business.board.dto;
+package com.ureca.picky_be.base.business.board.dto.boardDto;
+import com.ureca.picky_be.base.business.board.dto.contentDto.AddBoardContentReq;
+
 import java.util.List;
 
 public record AddBoardReq(String boardContext, Long movieId, List<AddBoardContentReq> contents, boolean isSpoiler) {
