+++ src/main/java/com/ureca/picky_be/base/business/board/dto/boardDto/AddBoardResp.java
@@ -1,0 +1,5 @@
+package com.ureca.picky_be.base.business.board.dto.boardDto;
+
+public record AddBoardResp() {
+
+}
