+++ src/main/java/com/ureca/picky_be/base/business/board/dto/boardDto/GetMovieRelatedBoardsResp.java
@@ -1,0 +1,4 @@
+package com.ureca.picky_be.base.business.board.dto.boardDto;
+
+public record GetMovieRelatedBoardsResp() {
+}
