+++ src/main/java/com/ureca/picky_be/base/business/board/dto/commentDto/AddBoardCommentReq.java
@@ -1,0 +1,4 @@
+package com.ureca.picky_be.base.business.board.dto.commentDto;
+
+public record AddBoardCommentReq(String content) {
+}
