+++ src/main/java/com/ureca/picky_be/base/business/board/dto/commentDto/DeleteBoardCommentsResp.java
@@ -1,0 +1,4 @@
+package com.ureca.picky_be.base.business.board.dto.commentDto;
+
+public record DeleteBoardCommentsResp() {
+}
