+++ src/main/java/com/ureca/picky_be/base/business/board/dto/contentDto/AddBoardContentReq.java
@@ -1,0 +1,4 @@
+package com.ureca.picky_be.base.business.board.dto.contentDto;
+
+public record AddBoardContentReq(String contentUrl, String type) {
+}
