+++ src/main/java/com/ureca/picky_be/base/business/board/dto/contentDto/GetBoardContentsResp.java
@@ -1,0 +1,5 @@
+package com.ureca.picky_be.base.business.board.dto.contentDto;
+
+public record GetBoardContentsResp(String contentUrl, String contentType) {
+
+}
