+++ src/main/java/com/ureca/picky_be/base/business/board/dto/likeDto/AddOrDeleteBoardLikeResp.java
@@ -1,0 +1,4 @@
+package com.ureca.picky_be.base.business.board.dto.likeDto;
+
+public record AddOrDeleteBoardLikeResp() {
+}
